requirements.txt
streamlit==1.39.0
httpx==0.27.2
PyMuPDF==1.24.14
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
